refactoring(tests) : server startup flow
