+++ service/aws/iam/accountalias_test.go
@@ -9,11 +9,90 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-
-	"github.com/rajasoun/aws-hub/service/aws/iam/iammock"
 )
 
-func TestGetAliases(t *testing.T) {
+// Client provides the API client to mock AWS operations
+type MockAccountAliases struct {
+	mock.Mock
+}
+
+/**
+* Mock using testify Framework
+ */
+
+// List Account Aliases Mock
+func (mockFunc *MockAccountAliases) ListAccountAliases(ctx context.Context,
+	params *iam.ListAccountAliasesInput,
+	optFns ...func(*iam.Options)) (*iam.ListAccountAliasesOutput, error) {
+	// Call Mock Function call
+	// The Function to be Called and Result will be Injected in Test
+	result := mockFunc.Called(ctx, params, optFns)
+	//Return Result On Error
+	if result.Get(0) == nil {
+		return nil, result.Error(1)
+	}
+	//Return Result If No Error
+	return result.Get(0).(*iam.ListAccountAliasesOutput), result.Error(1)
+}
+
+func TestListAccountAliases(t *testing.T) {
+	assert := assert.New(t)
+	t.Parallel()
+	t.Run("Check ListAccountAliases via Mocking Framework ", func(t *testing.T) {
+		//client := new(iammock.MockClient)
+		client := new(MockAccountAliases)
+		var testAlias string = "aws-test-account-alias"
+		aliases := []string{testAlias}
+		expectedOutput := &iam.ListAccountAliasesOutput{
+			AccountAliases: aliases,
+		}
+		// Inject Mock Function to be Called along with Resturn values as Parameter
+		client.
+			On("ListAccountAliases", mock.Anything, mock.Anything, mock.Anything).
+			Return(expectedOutput, nil)
+		got, err := GetAliases(client)
+		assert.NoError(err, "expect no error, got %v", err)
+		assert.Equal(testAlias, got.List[0], "got GetAliases = %v, want = %v", got.List[0], testAlias)
+	})
+	t.Run("Check GetUserCount returns err with Empty aws.Config{}", func(t *testing.T) {
+		emptyCfg := aws.Config{}
+		noOpClient := iam.NewFromConfig(emptyCfg)
+		_, err := GetAliases(noOpClient)
+		assert.Error(err, "err = %v, want = nil", err)
+	})
+}
+
+/**
+* Mock via manual creation - Just For Reference
+ */
+
+//Mock Function
+type MockListAccountAliasesAPIClient func(ctx context.Context,
+	params *iam.ListAccountAliasesInput,
+	optFns ...func(*iam.Options)) (*iam.ListAccountAliasesOutput, error)
+
+// Implement AWS IAM ListAccountAliases Interface with mock reciever
+func (mock MockListAccountAliasesAPIClient) ListAccountAliases(ctx context.Context,
+	params *iam.ListAccountAliasesInput,
+	optFns ...func(*iam.Options)) (*iam.ListAccountAliasesOutput, error) {
+	return mock(ctx, params, optFns...)
+}
+
+func (mock MockAccountAliases) NewClient() IAMListAccountAliasesAPIClient {
+	fn := func(ctx context.Context,
+		params *iam.ListAccountAliasesInput,
+		optFns ...func(*iam.Options)) (*iam.ListAccountAliasesOutput, error) {
+		var testAlias string = "aws-test-account-alias"
+		aliases := []string{testAlias}
+		result := &iam.ListAccountAliasesOutput{
+			AccountAliases: aliases,
+		}
+		return result, nil
+	}
+	client := MockListAccountAliasesAPIClient(fn)
+	return client
+}
+func TestGetAliasesviaHandMadeMock(t *testing.T) {
 	var testAlias string = "aws-test-account-alias"
 	assert := assert.New(t)
 	t.Parallel()
@@ -27,54 +106,11 @@ func TestGetAliases(t *testing.T) {
 
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
-			mock := iammock.MockAccountAliases{}
+			mock := MockAccountAliases{}
 			client := mock.NewClient()
 			got, err := GetAliases(client)
 			assert.NoError(err, "expect no error, got %v", err)
 			assert.Equal(tt.want, got.List[0], "got GetUserCount = %v, want = %v", got.List[0], tt.want)
 		})
 	}
-	t.Run("Check GetUserCount returns err with Empty aws.Config{}", func(t *testing.T) {
-		emptyCfg := aws.Config{}
-		noOpClient := iam.NewFromConfig(emptyCfg) //mock.NewMockClient(emptyCfg)
-		_, err := GetAliases(noOpClient)
-		assert.Error(err, "err = %v, want = nil", err)
-	})
-}
-
-// Client provides the API client to mock AWS operations
-type MockIt struct {
-	mock.Mock
-}
-
-// List Account Aliases Mock
-func (c *MockIt) ListAccountAliases(ctx context.Context,
-	params *iam.ListAccountAliasesInput,
-	optFns ...func(*iam.Options)) (*iam.ListAccountAliasesOutput, error) {
-	args := c.Called(ctx, params, optFns)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*iam.ListAccountAliasesOutput), args.Error(1)
-}
-
-func TestListAccountAliasesViaMockFramework(t *testing.T) {
-	assert := assert.New(t)
-	t.Parallel()
-	t.Run("Check ListAccountAliases via Mocking Framework ", func(t *testing.T) {
-		//client := new(iammock.MockClient)
-		client := new(MockIt)
-		var testAlias string = "aws-test-account-alias"
-		aliases := []string{testAlias}
-		expectedOutput := &iam.ListAccountAliasesOutput{
-			AccountAliases: aliases,
-		}
-		//client.InjectFunctionMock(client, "ListAccountAliases", expectedOutput)
-		client.
-			On("ListAccountAliases", mock.Anything, mock.Anything, mock.Anything).
-			Return(expectedOutput, nil)
-		got, err := GetAliases(client)
-		assert.NoError(err, "expect no error, got %v", err)
-		assert.Equal(testAlias, got.List[0], "got GetAliases = %v, want = %v", got.List[0], testAlias)
-	})
 }
